Simple Guessing Game
This project will create a simple text-based game that allows the user to play against the computer. The game will be a simple guessing game where the user tries to guess a number between 1 and 100. The computer will randomly generate a number, and the user will have 10 guesses to guess the correct number. The game will keep track of the user score and display it at the end of the game.